working fill trade table on one item in security table

+++ moex_web/moex_web.Data/Repositories/TradeRepository.cs
@@ -51,8 +51,11 @@ namespace moex_web.Data.Repositories
         {
             var context = _context.GetContext();
             var trades = await context.Trades.Where(t => DateTime.Compare(t.TradeDate, DateTime.Parse(oldDate)) < 0).ToListAsync();
-            context.Trades.RemoveRange(trades);
-            context.SaveChanges();
+            if (trades != null)
+            {
+                context.Trades.RemoveRange(trades);
+                context.SaveChanges();
+            }
         }
     }
 }
